refactor: middleware are now executed with a middleware executor
