+++ src/router/router.ts
@@ -17,8 +17,7 @@ import {
 	Resolver,
 } from './controller/params/decorators.ts';
 import { trimSlash } from './utils.ts';
-import { DanetMiddleware, middlewareMetadataKey } from './middleware.ts';
-import { InjectableConstructor } from '../injector/injectable/constructor.ts';
+import { MiddlewareExecutor } from './middleware/executor.ts';
 
 // deno-lint-ignore no-explicit-any
 export type Callback = (...args: any[]) => unknown;
@@ -34,6 +33,7 @@ export class DanetRouter {
 		private guardExecutor: GuardExecutor = new GuardExecutor(injector),
 		private filterExecutor: FilterExecutor = new FilterExecutor(),
 		private viewRenderer: Renderer = new HandlebarRenderer(),
+		private middlewareExecutor: MiddlewareExecutor = new MiddlewareExecutor(injector),
 	) {
 	}
 	methodsMap = new Map([
@@ -97,7 +97,10 @@ export class DanetRouter {
 	) {
 		return async (context: HttpContext) => {
 			try {
-				await this.executeMiddlewares(context, Controller, ControllerMethod);
+				await this.middlewareExecutor.executeAllRelevantMiddlewares(
+					context,
+					Controller,
+					ControllerMethod);
 				const controllerInstance = await this.injector.get(
 					Controller,
 					context,
@@ -140,28 +143,6 @@ export class DanetRouter {
 				context.response.status = status;
 			}
 		};
-	}
-
-	private async executeMiddlewares(
-		context: HttpContext,
-		Controller: ControllerConstructor,
-		ControllerMethod: Callback) {
-		const controllerMiddlewares: InjectableConstructor[] = MetadataHelper.getMetadata(middlewareMetadataKey, Controller);
-		if (controllerMiddlewares) {
-			await this.injector.registerInjectables(controllerMiddlewares);
-			for (const middlewareConstructor of controllerMiddlewares) {
-				const middlewareInstance: DanetMiddleware = await this.injector.get<DanetMiddleware>(middlewareConstructor as Constructor<DanetMiddleware>);
-				await middlewareInstance.action(context);
-			}
-		}
-		const methodMiddlewares: InjectableConstructor[] = MetadataHelper.getMetadata(middlewareMetadataKey, ControllerMethod);
-		if (methodMiddlewares) {
-			await this.injector.registerInjectables(methodMiddlewares);
-			for (const middlewareConstructor of methodMiddlewares) {
-				const middlewareInstance: DanetMiddleware = await this.injector.get<DanetMiddleware>(middlewareConstructor as Constructor<DanetMiddleware>);
-				await middlewareInstance.action(context);
-			}
-		}
 	}
 
 	private async sendResponse(
